# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sms-Sender API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "description": "Login with sanctum guard",
                "operationId": "apiLogin",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Users email",
                        "required": true,
                        "schema": {
                            "format": "email"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Users password",
                        "required": true,
                        "schema": {
                            "format": "password"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "",
                                            "type": "status",
                                            "default": "true"
                                        },
                                        "message": {
                                            "description": "",
                                            "type": "message",
                                            "default": "User logged in successfully"
                                        },
                                        "token": {
                                            "description": "",
                                            "type": "token",
                                            "default": "1|4tRvz5YdLzZP1k50IpSeqhlyAvXEp150OJYNFEFV"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "",
                                            "type": "status",
                                            "default": "false"
                                        },
                                        "message": {
                                            "description": "",
                                            "type": "message",
                                            "default": "Email & Password does not match with our record."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "email/password": {
                                            "description": "",
                                            "type": "message",
                                            "default": "The email/password  field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/send-sms/toky-digital": {
            "post": {
                "tags": [
                    "Send Sms"
                ],
                "summary": "Send Sms",
                "description": "Routes for sending sms",
                "operationId": "sendSmsTokyDigital",
                "parameters": [
                    {
                        "name": "sender",
                        "in": "query",
                        "description": "The one who sends sms",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "number",
                        "in": "query",
                        "description": "Number to whom is sms sent",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "message",
                        "in": "query",
                        "description": "Message we want to send",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "campaign",
                        "in": "query",
                        "description": "Campaign",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sms_type",
                        "in": "query",
                        "description": "Type of sms",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "contract_type",
                        "in": "query",
                        "description": "Type of contract",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Country where we send sms, one of : IT,UK,FR,USA,DE,ZV",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "provider",
                        "in": "query",
                        "description": "Provider which sends sms : tokyDigital",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "",
                                            "type": "message",
                                            "default": "Successfully sent"
                                        },
                                        "status": {
                                            "description": "",
                                            "type": "status",
                                            "default": "OK"
                                        },
                                        "full_status": {
                                            "description": "",
                                            "type": "full_status",
                                            "default": "+OK \\n Sent=1 \\n Errors=0 \\n Credit=1 \\n ITA=0 \\n FRA=48 \\n GBR=32032 \\n DEU=50 \\n USA=50 \\n."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "",
                                            "type": "message",
                                            "default": "Message not sent"
                                        },
                                        "status": {
                                            "description": "",
                                            "type": "status",
                                            "default": "ER"
                                        },
                                        "full_status": {
                                            "description": "",
                                            "type": "full_status",
                                            "default": "-ER The (field name) field is required. "
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/send-sms/gsm-box": {
            "post": {
                "tags": [
                    "Send Sms"
                ],
                "summary": "Send Sms through gsm box",
                "description": "Routes for sending sms",
                "operationId": "sendSmsGsmBox",
                "parameters": [
                    {
                        "name": "sender",
                        "in": "query",
                        "description": "The one who sends sms",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "number",
                        "in": "query",
                        "description": "Number to whom is sms sent",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "message",
                        "in": "query",
                        "description": "Message we want to send",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "campaign",
                        "in": "query",
                        "description": "Campaign",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sms_type",
                        "in": "query",
                        "description": "Type of sms",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "contract_type",
                        "in": "query",
                        "description": "Type of contract",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "Country where sms is sent",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "provider",
                        "in": "query",
                        "description": "Provider which sends sms",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "port",
                        "in": "query",
                        "description": "Port used to send this sms",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "",
                                            "type": "message",
                                            "default": "Successfully sent"
                                        },
                                        "status": {
                                            "description": "",
                                            "type": "status",
                                            "default": "OK"
                                        },
                                        "full_status": {
                                            "description": "",
                                            "type": "full_status",
                                            "default": "{\\'code\\':0, \\'reason\\':\\'OK\\'}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "country": {
                                            "description": "",
                                            "type": "country",
                                            "default": "The selected country is invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}